+++ airo-camera-toolkit/airo_camera_toolkit/image_converter.py
