+++ airo-camera-toolkit/test/test_image_converter.py
@@ -1,5 +1,10 @@
 import numpy as np
-from airo_camera_toolkit.utils import ImageConverter, is_float_image_array, is_image_array, is_int_image_array
+from airo_camera_toolkit.image_converter import (
+    ImageConverter,
+    is_float_image_array,
+    is_image_array,
+    is_int_image_array,
+)
 
 
 def test_image_format_checks():
